+++ packages/docs/src/components/PageFrame.astro
@@ -13,4 +13,5 @@ import Default from "@astrojs/starlight/components/PageFrame.astro";
   import "./panzoom.ts";
   import "./vizdom.ts";
   import "./graphviz.ts";
+  import "./d2.ts";
 </script>
